+++ Runtime/Scripts/Core/Types/LazySingleton.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿// Copyright (c) 2022 Jonathan Lang
+
+using System;
 
 namespace Baracuda.Monitoring.Types
 {